requirements.txt
numpy==1.24.3
pandas==1.5.3
tqdm==4.65.0
pyarrow
fastparquet

airflow
pyspark
sqlalchemy
psycopg2-binary

polars

black